requirements.txt
requests
schedule
pytz
feedparser
beautifulsoup4
python-dotenv
google-genai
